requirements.txt
chess~=1.11
PyYAML==6.0.1
requests~=2.32
backoff~=2.2
rich~=13.9
